Break out Line Chart specific logic into Line Chart component from Chart main component
